requirements.txt
coverage==6.2
cycler==0.11.0
kiwisolver==1.3.1
matplotlib==3.3.3
nose2==0.11.0
numpy==1.19.5
Pillow==8.4.0
pyparsing==3.0.7
python-dateutil==2.8.2
PyYAML==6.0
scipy==1.5.4
serpentTools==0.9.4
six==1.16.0
